Fix Python 3 support
When executing this module in Python 3 the following exception raises
"TypeError: Object of type map is not JSON serializable". In Python 3,
the map built-in function is lazy evaluated and this breaks the JSON
encoder.

This commit attempts to solve the problem described above. The "map()"
call is wrapped with a "list()" in order to evaluate the map.